apps/llm-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ONNX Runtime GenAI (lightweight, production-ready)
onnxruntime-genai>=0.3.0

# Optional: Full transformers stack for app_onnx.py
transformers>=4.40.0
optimum[onnxruntime]>=1.17.0
onnxruntime>=1.17.0
torch>=2.1.0
